+++ attributes/cfn-awslogs.rb
@@ -2,7 +2,7 @@ default['cfn']['awslogs'].tap do |config|
   config['user'] = 'root'
   config['group'] = 'root'
   config['path'] = '/var/awslogs'
-  config['version'] = '1.4.0'
+  config['version'] = '1.4.5'
   config['region'] = nil
   config['streams'] = {}
   config['service'] = 'awslogs'