+++ metadata.rb
@@ -6,11 +6,11 @@ description      'Chef integration with AWS cloudformation'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 issues_url       'https://github.com/JonathanSerafini/chef-cfn/issues'
 source_url       'https://github.com/JonathanSerafini/chef-cfn'
-version          '2.1.1'
+version          '3.1.0'
 
 depends 'chef_handler'
 depends 'ohai', '>= 4.0.0'
-depends 'python', '>= 1.4.0'
+depends 'poise-python', '~> 1.7.0'
 
 #
 # Attributes