+++ recipes/awslogs.rb
@@ -12,14 +12,21 @@ end
 
 python_virtualenv node['cfn']['awslogs']['path'] do
   action :create
-  owner node['cfn']['awslogs']['user']
+  user node['cfn']['awslogs']['user']
   group node['cfn']['awslogs']['group']
+  get_pip_url 'https://bootstrap.pypa.io/pip/2.7/get-pip.py'  
+  pip_version true
+  setuptools_version true
 end
 
-python_pip 'awscli-cwlogs' do
+python_execute "-m pip install setuptools" do
   virtualenv node['cfn']['awslogs']['path']
-  version node['cfn']['awslogs']['version']
 end
+
+python_execute "-m pip install awscli-cwlogs" do
+  virtualenv node['cfn']['awslogs']['path']
+end
+
 
 template "#{node['cfn']['awslogs']['path']}/etc/aws.conf" do
   owner 'root'