+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 chef\_cfn changelog
 ===================
 
+v3.1.0
+------
+* Found a way to install python2 using poise-python
+
+v3.0.0
+------
+* Replace python with poise-python and version bump to Python 3 but limiting to PIP v18.0
+
+v2.1.2
+------
+* Version-locked aws-ruby-sdk to '~> v2.10.98'
+
 v2.1.1
 ------
 * Add missing cloudwatch configuration attributes